LayoutTests/imported/w3c/web-platform-tests/webrtc/interfaces-expected.txt

CONSOLE MESSAGE: line 251: callback not yet supported
CONSOLE MESSAGE: line 251: callback not yet supported
CONSOLE MESSAGE: line 251: callback not yet supported
CONSOLE MESSAGE: line 251: callback not yet supported

PASS Test driver 
FAIL MediaStreamTrack interface: attribute isolated assert_true: The prototype object must have a property "isolated" expected true got false
FAIL MediaStreamTrack interface: attribute onisolationchange assert_true: The prototype object must have a property "onisolationchange" expected true got false
FAIL RTCPeerConnection interface: existence and properties of interface object undefined is not an object (evaluating 'this.array
                     .members[this.base]
                     .has_extended_attribute')
PASS RTCPeerConnection interface object length 
PASS RTCPeerConnection interface object name 
FAIL RTCPeerConnection interface: existence and properties of interface prototype object undefined is not an object (evaluating 'this.array
                         .members[inherit_interface]
                         .has_extended_attribute')
PASS RTCPeerConnection interface: existence and properties of interface prototype object's "constructor" property 
PASS RTCPeerConnection interface: operation createOffer(RTCOfferOptions) 
PASS RTCPeerConnection interface: operation createAnswer(RTCAnswerOptions) 
FAIL RTCPeerConnection interface: operation setLocalDescription(RTCSessionDescriptionInit) assert_equals: property has wrong .length expected 1 but got 0
PASS RTCPeerConnection interface: attribute localDescription 
PASS RTCPeerConnection interface: attribute currentLocalDescription 
PASS RTCPeerConnection interface: attribute pendingLocalDescription 
FAIL RTCPeerConnection interface: operation setRemoteDescription(RTCSessionDescriptionInit) assert_equals: property has wrong .length expected 1 but got 0
PASS RTCPeerConnection interface: attribute remoteDescription 
PASS RTCPeerConnection interface: attribute currentRemoteDescription 
PASS RTCPeerConnection interface: attribute pendingRemoteDescription 
PASS RTCPeerConnection interface: operation addIceCandidate([object Object],[object Object]) 
PASS RTCPeerConnection interface: attribute signalingState 
PASS RTCPeerConnection interface: attribute iceGatheringState 
PASS RTCPeerConnection interface: attribute iceConnectionState 
PASS RTCPeerConnection interface: attribute connectionState 
FAIL RTCPeerConnection interface: attribute canTrickleIceCandidates assert_true: The prototype object must have a property "canTrickleIceCandidates" expected true got false
FAIL RTCPeerConnection interface: operation getDefaultIceServers() assert_own_property: interface object missing static operation expected property "getDefaultIceServers" missing
PASS RTCPeerConnection interface: operation getConfiguration() 
PASS RTCPeerConnection interface: operation setConfiguration(RTCConfiguration) 
PASS RTCPeerConnection interface: operation close() 
PASS RTCPeerConnection interface: attribute onnegotiationneeded 
PASS RTCPeerConnection interface: attribute onicecandidate 
FAIL RTCPeerConnection interface: attribute onicecandidateerror assert_true: The prototype object must have a property "onicecandidateerror" expected true got false
PASS RTCPeerConnection interface: attribute onsignalingstatechange 
PASS RTCPeerConnection interface: attribute oniceconnectionstatechange 
PASS RTCPeerConnection interface: attribute onicegatheringstatechange 
PASS RTCPeerConnection interface: attribute onconnectionstatechange 
FAIL RTCPeerConnection interface: attribute onfingerprintfailure assert_true: The prototype object must have a property "onfingerprintfailure" expected true got false
PASS RTCPeerConnection interface: operation createOffer(RTCSessionDescriptionCallback,RTCPeerConnectionErrorCallback,RTCOfferOptions) 
FAIL RTCPeerConnection interface: operation setLocalDescription(RTCSessionDescriptionInit,VoidFunction,RTCPeerConnectionErrorCallback) assert_equals: property has wrong .length expected 1 but got 0
PASS RTCPeerConnection interface: operation createAnswer(RTCSessionDescriptionCallback,RTCPeerConnectionErrorCallback) 
FAIL RTCPeerConnection interface: operation setRemoteDescription(RTCSessionDescriptionInit,VoidFunction,RTCPeerConnectionErrorCallback) assert_equals: property has wrong .length expected 1 but got 0
PASS RTCPeerConnection interface: operation addIceCandidate([object Object],[object Object],VoidFunction,RTCPeerConnectionErrorCallback) 
FAIL RTCPeerConnection interface: operation generateCertificate(AlgorithmIdentifier) assert_own_property: interface object missing static operation expected property "generateCertificate" missing
PASS RTCPeerConnection interface: operation getSenders() 
PASS RTCPeerConnection interface: operation getReceivers() 
PASS RTCPeerConnection interface: operation getTransceivers() 
PASS RTCPeerConnection interface: operation addTrack(MediaStreamTrack,MediaStream) 
PASS RTCPeerConnection interface: operation removeTrack(RTCRtpSender) 
PASS RTCPeerConnection interface: operation addTransceiver([object Object],[object Object],RTCRtpTransceiverInit) 
PASS RTCPeerConnection interface: attribute ontrack 
FAIL RTCPeerConnection interface: attribute sctp assert_true: The prototype object must have a property "sctp" expected true got false
PASS RTCPeerConnection interface: operation createDataChannel(USVString,RTCDataChannelInit) 
PASS RTCPeerConnection interface: attribute ondatachannel 
PASS RTCPeerConnection interface: operation getStats(MediaStreamTrack) 
FAIL RTCPeerConnection interface: operation setIdentityProvider(DOMString,RTCIdentityProviderOptions) assert_own_property: interface prototype object missing non-static operation expected property "setIdentityProvider" missing
FAIL RTCPeerConnection interface: operation getIdentityAssertion() assert_own_property: interface prototype object missing non-static operation expected property "getIdentityAssertion" missing
FAIL RTCPeerConnection interface: attribute peerIdentity assert_true: The prototype object must have a property "peerIdentity" expected true got false
FAIL RTCPeerConnection interface: attribute idpLoginUrl assert_true: The prototype object must have a property "idpLoginUrl" expected true got false
FAIL RTCPeerConnection interface: attribute idpErrorInfo assert_true: The prototype object must have a property "idpErrorInfo" expected true got false
PASS RTCPeerConnection interface: new RTCPeerConnection must inherit property "createOffer" with the proper type (0) 
PASS RTCPeerConnection interface: calling createOffer(RTCOfferOptions) on new RTCPeerConnection with too few arguments must throw TypeError 
PASS RTCPeerConnection interface: new RTCPeerConnection must inherit property "createAnswer" with the proper type (1) 
PASS RTCPeerConnection interface: calling createAnswer(RTCAnswerOptions) on new RTCPeerConnection with too few arguments must throw TypeError 
PASS RTCPeerConnection interface: new RTCPeerConnection must inherit property "setLocalDescription" with the proper type (2) 
PASS RTCPeerConnection interface: calling setLocalDescription(RTCSessionDescriptionInit) on new RTCPeerConnection with too few arguments must throw TypeError 
PASS RTCPeerConnection interface: new RTCPeerConnection must inherit property "localDescription" with the proper type (3) 
PASS RTCPeerConnection interface: new RTCPeerConnection must inherit property "currentLocalDescription" with the proper type (4) 
PASS RTCPeerConnection interface: new RTCPeerConnection must inherit property "pendingLocalDescription" with the proper type (5) 
PASS RTCPeerConnection interface: new RTCPeerConnection must inherit property "setRemoteDescription" with the proper type (6) 
PASS RTCPeerConnection interface: calling setRemoteDescription(RTCSessionDescriptionInit) on new RTCPeerConnection with too few arguments must throw TypeError 
PASS RTCPeerConnection interface: new RTCPeerConnection must inherit property "remoteDescription" with the proper type (7) 
PASS RTCPeerConnection interface: new RTCPeerConnection must inherit property "currentRemoteDescription" with the proper type (8) 
PASS RTCPeerConnection interface: new RTCPeerConnection must inherit property "pendingRemoteDescription" with the proper type (9) 
PASS RTCPeerConnection interface: new RTCPeerConnection must inherit property "addIceCandidate" with the proper type (10) 
PASS RTCPeerConnection interface: calling addIceCandidate([object Object],[object Object]) on new RTCPeerConnection with too few arguments must throw TypeError 
PASS RTCPeerConnection interface: new RTCPeerConnection must inherit property "signalingState" with the proper type (11) 
PASS RTCPeerConnection interface: new RTCPeerConnection must inherit property "iceGatheringState" with the proper type (12) 
PASS RTCPeerConnection interface: new RTCPeerConnection must inherit property "iceConnectionState" with the proper type (13) 
PASS RTCPeerConnection interface: new RTCPeerConnection must inherit property "connectionState" with the proper type (14) 
FAIL RTCPeerConnection interface: new RTCPeerConnection must inherit property "canTrickleIceCandidates" with the proper type (15) assert_inherits: property "canTrickleIceCandidates" not found in prototype chain
PASS RTCPeerConnection interface: new RTCPeerConnection must inherit property "getDefaultIceServers" with the proper type (16) 
PASS RTCPeerConnection interface: new RTCPeerConnection must inherit property "getConfiguration" with the proper type (17) 
PASS RTCPeerConnection interface: new RTCPeerConnection must inherit property "setConfiguration" with the proper type (18) 
PASS RTCPeerConnection interface: calling setConfiguration(RTCConfiguration) on new RTCPeerConnection with too few arguments must throw TypeError 
PASS RTCPeerConnection interface: new RTCPeerConnection must inherit property "close" with the proper type (19) 
PASS RTCPeerConnection interface: new RTCPeerConnection must inherit property "onnegotiationneeded" with the proper type (20) 
PASS RTCPeerConnection interface: new RTCPeerConnection must inherit property "onicecandidate" with the proper type (21) 
FAIL RTCPeerConnection interface: new RTCPeerConnection must inherit property "onicecandidateerror" with the proper type (22) assert_inherits: property "onicecandidateerror" not found in prototype chain
PASS RTCPeerConnection interface: new RTCPeerConnection must inherit property "onsignalingstatechange" with the proper type (23) 
PASS RTCPeerConnection interface: new RTCPeerConnection must inherit property "oniceconnectionstatechange" with the proper type (24) 
PASS RTCPeerConnection interface: new RTCPeerConnection must inherit property "onicegatheringstatechange" with the proper type (25) 
PASS RTCPeerConnection interface: new RTCPeerConnection must inherit property "onconnectionstatechange" with the proper type (26) 
FAIL RTCPeerConnection interface: new RTCPeerConnection must inherit property "onfingerprintfailure" with the proper type (27) assert_inherits: property "onfingerprintfailure" not found in prototype chain
PASS RTCPeerConnection interface: new RTCPeerConnection must inherit property "createOffer" with the proper type (28) 
PASS RTCPeerConnection interface: calling createOffer(RTCSessionDescriptionCallback,RTCPeerConnectionErrorCallback,RTCOfferOptions) on new RTCPeerConnection with too few arguments must throw TypeError 
PASS RTCPeerConnection interface: new RTCPeerConnection must inherit property "setLocalDescription" with the proper type (29) 
PASS RTCPeerConnection interface: calling setLocalDescription(RTCSessionDescriptionInit,VoidFunction,RTCPeerConnectionErrorCallback) on new RTCPeerConnection with too few arguments must throw TypeError 
PASS RTCPeerConnection interface: new RTCPeerConnection must inherit property "createAnswer" with the proper type (30) 
PASS RTCPeerConnection interface: calling createAnswer(RTCSessionDescriptionCallback,RTCPeerConnectionErrorCallback) on new RTCPeerConnection with too few arguments must throw TypeError 
PASS RTCPeerConnection interface: new RTCPeerConnection must inherit property "setRemoteDescription" with the proper type (31) 
PASS RTCPeerConnection interface: calling setRemoteDescription(RTCSessionDescriptionInit,VoidFunction,RTCPeerConnectionErrorCallback) on new RTCPeerConnection with too few arguments must throw TypeError 
PASS RTCPeerConnection interface: new RTCPeerConnection must inherit property "addIceCandidate" with the proper type (32) 
PASS RTCPeerConnection interface: calling addIceCandidate([object Object],[object Object],VoidFunction,RTCPeerConnectionErrorCallback) on new RTCPeerConnection with too few arguments must throw TypeError 
PASS RTCPeerConnection interface: new RTCPeerConnection must inherit property "generateCertificate" with the proper type (33) 
FAIL RTCPeerConnection interface: calling generateCertificate(AlgorithmIdentifier) on new RTCPeerConnection with too few arguments must throw TypeError assert_unreached: Throws "TypeError: undefined is not an object (evaluating 'fn.apply')" instead of rejecting promise Reached unreachable code
PASS RTCPeerConnection interface: new RTCPeerConnection must inherit property "getSenders" with the proper type (34) 
PASS RTCPeerConnection interface: new RTCPeerConnection must inherit property "getReceivers" with the proper type (35) 
PASS RTCPeerConnection interface: new RTCPeerConnection must inherit property "getTransceivers" with the proper type (36) 
PASS RTCPeerConnection interface: new RTCPeerConnection must inherit property "addTrack" with the proper type (37) 
PASS RTCPeerConnection interface: calling addTrack(MediaStreamTrack,MediaStream) on new RTCPeerConnection with too few arguments must throw TypeError 
PASS RTCPeerConnection interface: new RTCPeerConnection must inherit property "removeTrack" with the proper type (38) 
PASS RTCPeerConnection interface: calling removeTrack(RTCRtpSender) on new RTCPeerConnection with too few arguments must throw TypeError 
PASS RTCPeerConnection interface: new RTCPeerConnection must inherit property "addTransceiver" with the proper type (39) 
PASS RTCPeerConnection interface: calling addTransceiver([object Object],[object Object],RTCRtpTransceiverInit) on new RTCPeerConnection with too few arguments must throw TypeError 
PASS RTCPeerConnection interface: new RTCPeerConnection must inherit property "ontrack" with the proper type (40) 
FAIL RTCPeerConnection interface: new RTCPeerConnection must inherit property "sctp" with the proper type (41) assert_inherits: property "sctp" not found in prototype chain
PASS RTCPeerConnection interface: new RTCPeerConnection must inherit property "createDataChannel" with the proper type (42) 
PASS RTCPeerConnection interface: calling createDataChannel(USVString,RTCDataChannelInit) on new RTCPeerConnection with too few arguments must throw TypeError 
PASS RTCPeerConnection interface: new RTCPeerConnection must inherit property "ondatachannel" with the proper type (43) 
PASS RTCPeerConnection interface: new RTCPeerConnection must inherit property "getStats" with the proper type (44) 
PASS RTCPeerConnection interface: calling getStats(MediaStreamTrack) on new RTCPeerConnection with too few arguments must throw TypeError 
FAIL RTCPeerConnection interface: new RTCPeerConnection must inherit property "setIdentityProvider" with the proper type (45) assert_inherits: property "setIdentityProvider" not found in prototype chain
FAIL RTCPeerConnection interface: calling setIdentityProvider(DOMString,RTCIdentityProviderOptions) on new RTCPeerConnection with too few arguments must throw TypeError assert_inherits: property "setIdentityProvider" not found in prototype chain
FAIL RTCPeerConnection interface: new RTCPeerConnection must inherit property "getIdentityAssertion" with the proper type (46) assert_inherits: property "getIdentityAssertion" not found in prototype chain
FAIL RTCPeerConnection interface: new RTCPeerConnection must inherit property "peerIdentity" with the proper type (47) assert_inherits: property "peerIdentity" not found in prototype chain
FAIL RTCPeerConnection interface: new RTCPeerConnection must inherit property "idpLoginUrl" with the proper type (48) assert_inherits: property "idpLoginUrl" not found in prototype chain
FAIL RTCPeerConnection interface: new RTCPeerConnection must inherit property "idpErrorInfo" with the proper type (49) assert_inherits: property "idpErrorInfo" not found in prototype chain
PASS RTCSessionDescription interface: existence and properties of interface object 
PASS RTCSessionDescription interface object length 
PASS RTCSessionDescription interface object name 
PASS RTCSessionDescription interface: existence and properties of interface prototype object 
PASS RTCSessionDescription interface: existence and properties of interface prototype object's "constructor" property 
PASS RTCSessionDescription interface: attribute type 
PASS RTCSessionDescription interface: attribute sdp 
PASS RTCSessionDescription interface: new RTCSessionDescription({ type: "offer" }) must inherit property "type" with the proper type (0) 
PASS RTCSessionDescription interface: new RTCSessionDescription({ type: "offer" }) must inherit property "sdp" with the proper type (1) 
PASS RTCIceCandidate interface: existence and properties of interface object 
FAIL RTCIceCandidate interface object length assert_equals: wrong value for RTCIceCandidate.length expected 0 but got 1
PASS RTCIceCandidate interface object name 
PASS RTCIceCandidate interface: existence and properties of interface prototype object 
PASS RTCIceCandidate interface: existence and properties of interface prototype object's "constructor" property 
PASS RTCIceCandidate interface: attribute candidate 
PASS RTCIceCandidate interface: attribute sdpMid 
PASS RTCIceCandidate interface: attribute sdpMLineIndex 
FAIL RTCIceCandidate interface: attribute foundation assert_true: The prototype object must have a property "foundation" expected true got false
FAIL RTCIceCandidate interface: attribute component assert_true: The prototype object must have a property "component" expected true got false
FAIL RTCIceCandidate interface: attribute priority assert_true: The prototype object must have a property "priority" expected true got false
FAIL RTCIceCandidate interface: attribute ip assert_true: The prototype object must have a property "ip" expected true got false
FAIL RTCIceCandidate interface: attribute protocol assert_true: The prototype object must have a property "protocol" expected true got false
FAIL RTCIceCandidate interface: attribute port assert_true: The prototype object must have a property "port" expected true got false
FAIL RTCIceCandidate interface: attribute type assert_true: The prototype object must have a property "type" expected true got false
FAIL RTCIceCandidate interface: attribute tcpType assert_true: The prototype object must have a property "tcpType" expected true got false
FAIL RTCIceCandidate interface: attribute relatedAddress assert_true: The prototype object must have a property "relatedAddress" expected true got false
FAIL RTCIceCandidate interface: attribute relatedPort assert_true: The prototype object must have a property "relatedPort" expected true got false
FAIL RTCIceCandidate interface: attribute ufrag assert_true: The prototype object must have a property "ufrag" expected true got false
FAIL RTCIceCandidate interface: new RTCIceCandidate must inherit property "candidate" with the proper type (0) assert_equals: Unexpected exception when evaluating object expected null but got object "TypeError: Not enough arguments"
FAIL RTCIceCandidate interface: new RTCIceCandidate must inherit property "sdpMid" with the proper type (1) assert_equals: Unexpected exception when evaluating object expected null but got object "TypeError: Not enough arguments"
FAIL RTCIceCandidate interface: new RTCIceCandidate must inherit property "sdpMLineIndex" with the proper type (2) assert_equals: Unexpected exception when evaluating object expected null but got object "TypeError: Not enough arguments"
FAIL RTCIceCandidate interface: new RTCIceCandidate must inherit property "foundation" with the proper type (3) assert_equals: Unexpected exception when evaluating object expected null but got object "TypeError: Not enough arguments"
FAIL RTCIceCandidate interface: new RTCIceCandidate must inherit property "component" with the proper type (4) assert_equals: Unexpected exception when evaluating object expected null but got object "TypeError: Not enough arguments"
FAIL RTCIceCandidate interface: new RTCIceCandidate must inherit property "priority" with the proper type (5) assert_equals: Unexpected exception when evaluating object expected null but got object "TypeError: Not enough arguments"
FAIL RTCIceCandidate interface: new RTCIceCandidate must inherit property "ip" with the proper type (6) assert_equals: Unexpected exception when evaluating object expected null but got object "TypeError: Not enough arguments"
FAIL RTCIceCandidate interface: new RTCIceCandidate must inherit property "protocol" with the proper type (7) assert_equals: Unexpected exception when evaluating object expected null but got object "TypeError: Not enough arguments"
FAIL RTCIceCandidate interface: new RTCIceCandidate must inherit property "port" with the proper type (8) assert_equals: Unexpected exception when evaluating object expected null but got object "TypeError: Not enough arguments"
FAIL RTCIceCandidate interface: new RTCIceCandidate must inherit property "type" with the proper type (9) assert_equals: Unexpected exception when evaluating object expected null but got object "TypeError: Not enough arguments"
FAIL RTCIceCandidate interface: new RTCIceCandidate must inherit property "tcpType" with the proper type (10) assert_equals: Unexpected exception when evaluating object expected null but got object "TypeError: Not enough arguments"
FAIL RTCIceCandidate interface: new RTCIceCandidate must inherit property "relatedAddress" with the proper type (11) assert_equals: Unexpected exception when evaluating object expected null but got object "TypeError: Not enough arguments"
FAIL RTCIceCandidate interface: new RTCIceCandidate must inherit property "relatedPort" with the proper type (12) assert_equals: Unexpected exception when evaluating object expected null but got object "TypeError: Not enough arguments"
FAIL RTCIceCandidate interface: new RTCIceCandidate must inherit property "ufrag" with the proper type (13) assert_equals: Unexpected exception when evaluating object expected null but got object "TypeError: Not enough arguments"
FAIL RTCPeerConnectionIceEvent interface: existence and properties of interface object assert_own_property: self does not have own property "RTCPeerConnectionIceEvent" expected property "RTCPeerConnectionIceEvent" missing
FAIL RTCPeerConnectionIceEvent interface object length assert_own_property: self does not have own property "RTCPeerConnectionIceEvent" expected property "RTCPeerConnectionIceEvent" missing
FAIL RTCPeerConnectionIceEvent interface object name assert_own_property: self does not have own property "RTCPeerConnectionIceEvent" expected property "RTCPeerConnectionIceEvent" missing
FAIL RTCPeerConnectionIceEvent interface: existence and properties of interface prototype object assert_own_property: self does not have own property "RTCPeerConnectionIceEvent" expected property "RTCPeerConnectionIceEvent" missing
FAIL RTCPeerConnectionIceEvent interface: existence and properties of interface prototype object's "constructor" property assert_own_property: self does not have own property "RTCPeerConnectionIceEvent" expected property "RTCPeerConnectionIceEvent" missing
FAIL RTCPeerConnectionIceEvent interface: attribute candidate assert_own_property: self does not have own property "RTCPeerConnectionIceEvent" expected property "RTCPeerConnectionIceEvent" missing
FAIL RTCPeerConnectionIceEvent interface: attribute url assert_own_property: self does not have own property "RTCPeerConnectionIceEvent" expected property "RTCPeerConnectionIceEvent" missing
FAIL RTCPeerConnectionIceEvent interface: new RTCPeerConnectionIceEvent("ice") must inherit property "candidate" with the proper type (0) assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: Can't find variable: RTCPeerConnectionIceEvent"
FAIL RTCPeerConnectionIceEvent interface: new RTCPeerConnectionIceEvent("ice") must inherit property "url" with the proper type (1) assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: Can't find variable: RTCPeerConnectionIceEvent"
FAIL RTCPeerConnectionIceErrorEvent interface: existence and properties of interface object assert_own_property: self does not have own property "RTCPeerConnectionIceErrorEvent" expected property "RTCPeerConnectionIceErrorEvent" missing
FAIL RTCPeerConnectionIceErrorEvent interface object length assert_own_property: self does not have own property "RTCPeerConnectionIceErrorEvent" expected property "RTCPeerConnectionIceErrorEvent" missing
FAIL RTCPeerConnectionIceErrorEvent interface object name assert_own_property: self does not have own property "RTCPeerConnectionIceErrorEvent" expected property "RTCPeerConnectionIceErrorEvent" missing
FAIL RTCPeerConnectionIceErrorEvent interface: existence and properties of interface prototype object assert_own_property: self does not have own property "RTCPeerConnectionIceErrorEvent" expected property "RTCPeerConnectionIceErrorEvent" missing
FAIL RTCPeerConnectionIceErrorEvent interface: existence and properties of interface prototype object's "constructor" property assert_own_property: self does not have own property "RTCPeerConnectionIceErrorEvent" expected property "RTCPeerConnectionIceErrorEvent" missing
FAIL RTCPeerConnectionIceErrorEvent interface: attribute hostCandidate assert_own_property: self does not have own property "RTCPeerConnectionIceErrorEvent" expected property "RTCPeerConnectionIceErrorEvent" missing
FAIL RTCPeerConnectionIceErrorEvent interface: attribute url assert_own_property: self does not have own property "RTCPeerConnectionIceErrorEvent" expected property "RTCPeerConnectionIceErrorEvent" missing
FAIL RTCPeerConnectionIceErrorEvent interface: attribute errorCode assert_own_property: self does not have own property "RTCPeerConnectionIceErrorEvent" expected property "RTCPeerConnectionIceErrorEvent" missing
FAIL RTCPeerConnectionIceErrorEvent interface: attribute errorText assert_own_property: self does not have own property "RTCPeerConnectionIceErrorEvent" expected property "RTCPeerConnectionIceErrorEvent" missing
FAIL RTCPeerConnectionIceErrorEvent interface: new RTCPeerConnectionIceErrorEvent("ice-error", { errorCode: 701 }); must inherit property "hostCandidate" with the proper type (0) assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: Can't find variable: RTCPeerConnectionIceErrorEvent"
FAIL RTCPeerConnectionIceErrorEvent interface: new RTCPeerConnectionIceErrorEvent("ice-error", { errorCode: 701 }); must inherit property "url" with the proper type (1) assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: Can't find variable: RTCPeerConnectionIceErrorEvent"
FAIL RTCPeerConnectionIceErrorEvent interface: new RTCPeerConnectionIceErrorEvent("ice-error", { errorCode: 701 }); must inherit property "errorCode" with the proper type (2) assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: Can't find variable: RTCPeerConnectionIceErrorEvent"
FAIL RTCPeerConnectionIceErrorEvent interface: new RTCPeerConnectionIceErrorEvent("ice-error", { errorCode: 701 }); must inherit property "errorText" with the proper type (3) assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: Can't find variable: RTCPeerConnectionIceErrorEvent"
FAIL RTCCertificate interface: existence and properties of interface object assert_own_property: self does not have own property "RTCCertificate" expected property "RTCCertificate" missing
FAIL RTCCertificate interface object length assert_own_property: self does not have own property "RTCCertificate" expected property "RTCCertificate" missing
FAIL RTCCertificate interface object name assert_own_property: self does not have own property "RTCCertificate" expected property "RTCCertificate" missing
FAIL RTCCertificate interface: existence and properties of interface prototype object assert_own_property: self does not have own property "RTCCertificate" expected property "RTCCertificate" missing
FAIL RTCCertificate interface: existence and properties of interface prototype object's "constructor" property assert_own_property: self does not have own property "RTCCertificate" expected property "RTCCertificate" missing
FAIL RTCCertificate interface: attribute expires assert_own_property: self does not have own property "RTCCertificate" expected property "RTCCertificate" missing
FAIL RTCCertificate interface: operation getFingerprints() assert_own_property: self does not have own property "RTCCertificate" expected property "RTCCertificate" missing
FAIL RTCCertificate interface: operation getAlgorithm() assert_own_property: self does not have own property "RTCCertificate" expected property "RTCCertificate" missing
PASS RTCRtpSender interface: existence and properties of interface object 
PASS RTCRtpSender interface object length 
PASS RTCRtpSender interface object name 
PASS RTCRtpSender interface: existence and properties of interface prototype object 
PASS RTCRtpSender interface: existence and properties of interface prototype object's "constructor" property 
PASS RTCRtpSender interface: attribute track 
FAIL RTCRtpSender interface: attribute transport assert_true: The prototype object must have a property "transport" expected true got false
FAIL RTCRtpSender interface: attribute rtcpTransport assert_true: The prototype object must have a property "rtcpTransport" expected true got false
FAIL RTCRtpSender interface: operation getCapabilities(DOMString) assert_own_property: interface object missing static operation expected property "getCapabilities" missing
FAIL RTCRtpSender interface: operation setParameters(RTCRtpParameters) assert_own_property: interface prototype object missing non-static operation expected property "setParameters" missing
PASS RTCRtpSender interface: operation getParameters() 
PASS RTCRtpSender interface: operation replaceTrack(MediaStreamTrack) 
FAIL RTCRtpSender interface: operation getStats() assert_own_property: interface prototype object missing non-static operation expected property "getStats" missing
FAIL RTCRtpSender interface: attribute dtmf assert_true: The prototype object must have a property "dtmf" expected true got false
PASS RTCRtpReceiver interface: existence and properties of interface object 
PASS RTCRtpReceiver interface object length 
PASS RTCRtpReceiver interface object name 
PASS RTCRtpReceiver interface: existence and properties of interface prototype object 
PASS RTCRtpReceiver interface: existence and properties of interface prototype object's "constructor" property 
PASS RTCRtpReceiver interface: attribute track 
FAIL RTCRtpReceiver interface: attribute transport assert_true: The prototype object must have a property "transport" expected true got false
FAIL RTCRtpReceiver interface: attribute rtcpTransport assert_true: The prototype object must have a property "rtcpTransport" expected true got false
FAIL RTCRtpReceiver interface: operation getCapabilities(DOMString) assert_own_property: interface object missing static operation expected property "getCapabilities" missing
PASS RTCRtpReceiver interface: operation getParameters() 
FAIL RTCRtpReceiver interface: operation getContributingSources() assert_own_property: interface prototype object missing non-static operation expected property "getContributingSources" missing
FAIL RTCRtpReceiver interface: operation getSynchronizationSources() assert_own_property: interface prototype object missing non-static operation expected property "getSynchronizationSources" missing
FAIL RTCRtpReceiver interface: operation getStats() assert_own_property: interface prototype object missing non-static operation expected property "getStats" missing
FAIL RTCRtpContributingSource interface: existence and properties of interface object assert_own_property: self does not have own property "RTCRtpContributingSource" expected property "RTCRtpContributingSource" missing
FAIL RTCRtpContributingSource interface object length assert_own_property: self does not have own property "RTCRtpContributingSource" expected property "RTCRtpContributingSource" missing
FAIL RTCRtpContributingSource interface object name assert_own_property: self does not have own property "RTCRtpContributingSource" expected property "RTCRtpContributingSource" missing
FAIL RTCRtpContributingSource interface: existence and properties of interface prototype object assert_own_property: self does not have own property "RTCRtpContributingSource" expected property "RTCRtpContributingSource" missing
FAIL RTCRtpContributingSource interface: existence and properties of interface prototype object's "constructor" property assert_own_property: self does not have own property "RTCRtpContributingSource" expected property "RTCRtpContributingSource" missing
FAIL RTCRtpContributingSource interface: attribute timestamp assert_own_property: self does not have own property "RTCRtpContributingSource" expected property "RTCRtpContributingSource" missing
FAIL RTCRtpContributingSource interface: attribute source assert_own_property: self does not have own property "RTCRtpContributingSource" expected property "RTCRtpContributingSource" missing
FAIL RTCRtpContributingSource interface: attribute audioLevel assert_own_property: self does not have own property "RTCRtpContributingSource" expected property "RTCRtpContributingSource" missing
FAIL RTCRtpSynchronizationSource interface: existence and properties of interface object assert_own_property: self does not have own property "RTCRtpSynchronizationSource" expected property "RTCRtpSynchronizationSource" missing
FAIL RTCRtpSynchronizationSource interface object length assert_own_property: self does not have own property "RTCRtpSynchronizationSource" expected property "RTCRtpSynchronizationSource" missing
FAIL RTCRtpSynchronizationSource interface object name assert_own_property: self does not have own property "RTCRtpSynchronizationSource" expected property "RTCRtpSynchronizationSource" missing
FAIL RTCRtpSynchronizationSource interface: existence and properties of interface prototype object assert_own_property: self does not have own property "RTCRtpSynchronizationSource" expected property "RTCRtpSynchronizationSource" missing
FAIL RTCRtpSynchronizationSource interface: existence and properties of interface prototype object's "constructor" property assert_own_property: self does not have own property "RTCRtpSynchronizationSource" expected property "RTCRtpSynchronizationSource" missing
FAIL RTCRtpSynchronizationSource interface: attribute timestamp assert_own_property: self does not have own property "RTCRtpSynchronizationSource" expected property "RTCRtpSynchronizationSource" missing
FAIL RTCRtpSynchronizationSource interface: attribute source assert_own_property: self does not have own property "RTCRtpSynchronizationSource" expected property "RTCRtpSynchronizationSource" missing
FAIL RTCRtpSynchronizationSource interface: attribute audioLevel assert_own_property: self does not have own property "RTCRtpSynchronizationSource" expected property "RTCRtpSynchronizationSource" missing
FAIL RTCRtpSynchronizationSource interface: attribute voiceActivityFlag assert_own_property: self does not have own property "RTCRtpSynchronizationSource" expected property "RTCRtpSynchronizationSource" missing
PASS RTCRtpTransceiver interface: existence and properties of interface object 
PASS RTCRtpTransceiver interface object length 
PASS RTCRtpTransceiver interface object name 
PASS RTCRtpTransceiver interface: existence and properties of interface prototype object 
PASS RTCRtpTransceiver interface: existence and properties of interface prototype object's "constructor" property 
PASS RTCRtpTransceiver interface: attribute mid 
PASS RTCRtpTransceiver interface: attribute sender 
PASS RTCRtpTransceiver interface: attribute receiver 
PASS RTCRtpTransceiver interface: attribute stopped 
PASS RTCRtpTransceiver interface: attribute direction 
FAIL RTCRtpTransceiver interface: attribute currentDirection assert_true: The prototype object must have a property "currentDirection" expected true got false
PASS RTCRtpTransceiver interface: operation setDirection(RTCRtpTransceiverDirection) 
PASS RTCRtpTransceiver interface: operation stop() 
FAIL RTCRtpTransceiver interface: operation setCodecPreferences([object Object]) assert_own_property: interface prototype object missing non-static operation expected property "setCodecPreferences" missing
FAIL RTCDtlsTransport interface: existence and properties of interface object assert_own_property: self does not have own property "RTCDtlsTransport" expected property "RTCDtlsTransport" missing
FAIL RTCDtlsTransport interface object length assert_own_property: self does not have own property "RTCDtlsTransport" expected property "RTCDtlsTransport" missing
FAIL RTCDtlsTransport interface object name assert_own_property: self does not have own property "RTCDtlsTransport" expected property "RTCDtlsTransport" missing
FAIL RTCDtlsTransport interface: existence and properties of interface prototype object assert_own_property: self does not have own property "RTCDtlsTransport" expected property "RTCDtlsTransport" missing
FAIL RTCDtlsTransport interface: existence and properties of interface prototype object's "constructor" property assert_own_property: self does not have own property "RTCDtlsTransport" expected property "RTCDtlsTransport" missing
FAIL RTCDtlsTransport interface: attribute transport assert_own_property: self does not have own property "RTCDtlsTransport" expected property "RTCDtlsTransport" missing
FAIL RTCDtlsTransport interface: attribute state assert_own_property: self does not have own property "RTCDtlsTransport" expected property "RTCDtlsTransport" missing
FAIL RTCDtlsTransport interface: operation getRemoteCertificates() assert_own_property: self does not have own property "RTCDtlsTransport" expected property "RTCDtlsTransport" missing
FAIL RTCDtlsTransport interface: attribute onstatechange assert_own_property: self does not have own property "RTCDtlsTransport" expected property "RTCDtlsTransport" missing
PASS RTCIceTransport interface: existence and properties of interface object 
PASS RTCIceTransport interface object length 
PASS RTCIceTransport interface object name 
PASS RTCIceTransport interface: existence and properties of interface prototype object 
PASS RTCIceTransport interface: existence and properties of interface prototype object's "constructor" property 
FAIL RTCIceTransport interface: attribute role assert_true: The prototype object must have a property "role" expected true got false
FAIL RTCIceTransport interface: attribute component assert_true: The prototype object must have a property "component" expected true got false
PASS RTCIceTransport interface: attribute state 
PASS RTCIceTransport interface: attribute gatheringState 
FAIL RTCIceTransport interface: operation getLocalCandidates() assert_own_property: interface prototype object missing non-static operation expected property "getLocalCandidates" missing
FAIL RTCIceTransport interface: operation getRemoteCandidates() assert_own_property: interface prototype object missing non-static operation expected property "getRemoteCandidates" missing
FAIL RTCIceTransport interface: operation getSelectedCandidatePair() assert_own_property: interface prototype object missing non-static operation expected property "getSelectedCandidatePair" missing
FAIL RTCIceTransport interface: operation getLocalParameters() assert_own_property: interface prototype object missing non-static operation expected property "getLocalParameters" missing
FAIL RTCIceTransport interface: operation getRemoteParameters() assert_own_property: interface prototype object missing non-static operation expected property "getRemoteParameters" missing
FAIL RTCIceTransport interface: attribute onstatechange assert_true: The prototype object must have a property "onstatechange" expected true got false
FAIL RTCIceTransport interface: attribute ongatheringstatechange assert_true: The prototype object must have a property "ongatheringstatechange" expected true got false
FAIL RTCIceTransport interface: attribute onselectedcandidatepairchange assert_true: The prototype object must have a property "onselectedcandidatepairchange" expected true got false
FAIL RTCTrackEvent interface: existence and properties of interface object undefined is not an object (evaluating 'this.array
                     .members[this.base]
                     .has_extended_attribute')
PASS RTCTrackEvent interface object length 
PASS RTCTrackEvent interface object name 
FAIL RTCTrackEvent interface: existence and properties of interface prototype object undefined is not an object (evaluating 'this.array
                         .members[inherit_interface]
                         .has_extended_attribute')
PASS RTCTrackEvent interface: existence and properties of interface prototype object's "constructor" property 
PASS RTCTrackEvent interface: attribute receiver 
PASS RTCTrackEvent interface: attribute track 
PASS RTCTrackEvent interface: attribute streams 
PASS RTCTrackEvent interface: attribute transceiver 
FAIL RTCSctpTransport interface: existence and properties of interface object assert_own_property: self does not have own property "RTCSctpTransport" expected property "RTCSctpTransport" missing
FAIL RTCSctpTransport interface object length assert_own_property: self does not have own property "RTCSctpTransport" expected property "RTCSctpTransport" missing
FAIL RTCSctpTransport interface object name assert_own_property: self does not have own property "RTCSctpTransport" expected property "RTCSctpTransport" missing
FAIL RTCSctpTransport interface: existence and properties of interface prototype object assert_own_property: self does not have own property "RTCSctpTransport" expected property "RTCSctpTransport" missing
FAIL RTCSctpTransport interface: existence and properties of interface prototype object's "constructor" property assert_own_property: self does not have own property "RTCSctpTransport" expected property "RTCSctpTransport" missing
FAIL RTCSctpTransport interface: attribute transport assert_own_property: self does not have own property "RTCSctpTransport" expected property "RTCSctpTransport" missing
FAIL RTCSctpTransport interface: attribute maxMessageSize assert_own_property: self does not have own property "RTCSctpTransport" expected property "RTCSctpTransport" missing
FAIL RTCDataChannel interface: existence and properties of interface object undefined is not an object (evaluating 'this.array
                     .members[this.base]
                     .has_extended_attribute')
PASS RTCDataChannel interface object length 
PASS RTCDataChannel interface object name 
FAIL RTCDataChannel interface: existence and properties of interface prototype object undefined is not an object (evaluating 'this.array
                         .members[inherit_interface]
                         .has_extended_attribute')
PASS RTCDataChannel interface: existence and properties of interface prototype object's "constructor" property 
PASS RTCDataChannel interface: attribute label 
PASS RTCDataChannel interface: attribute ordered 
PASS RTCDataChannel interface: attribute maxPacketLifeTime 
PASS RTCDataChannel interface: attribute maxRetransmits 
PASS RTCDataChannel interface: attribute protocol 
PASS RTCDataChannel interface: attribute negotiated 
PASS RTCDataChannel interface: attribute id 
FAIL RTCDataChannel interface: attribute priority assert_true: The prototype object must have a property "priority" expected true got false
PASS RTCDataChannel interface: attribute readyState 
PASS RTCDataChannel interface: attribute bufferedAmount 
PASS RTCDataChannel interface: attribute bufferedAmountLowThreshold 
PASS RTCDataChannel interface: attribute onopen 
PASS RTCDataChannel interface: attribute onbufferedamountlow 
PASS RTCDataChannel interface: attribute onerror 
PASS RTCDataChannel interface: attribute onclose 
PASS RTCDataChannel interface: operation close() 
PASS RTCDataChannel interface: attribute onmessage 
PASS RTCDataChannel interface: attribute binaryType 
PASS RTCDataChannel interface: operation send(USVString) 
PASS RTCDataChannel interface: operation send(Blob) 
PASS RTCDataChannel interface: operation send(ArrayBuffer) 
PASS RTCDataChannel interface: operation send(ArrayBufferView) 
FAIL RTCDataChannelEvent interface: existence and properties of interface object undefined is not an object (evaluating 'this.array
                     .members[this.base]
                     .has_extended_attribute')
PASS RTCDataChannelEvent interface object length 
PASS RTCDataChannelEvent interface object name 
FAIL RTCDataChannelEvent interface: existence and properties of interface prototype object undefined is not an object (evaluating 'this.array
                         .members[inherit_interface]
                         .has_extended_attribute')
PASS RTCDataChannelEvent interface: existence and properties of interface prototype object's "constructor" property 
PASS RTCDataChannelEvent interface: attribute channel 
FAIL RTCDTMFSender interface: existence and properties of interface object assert_own_property: self does not have own property "RTCDTMFSender" expected property "RTCDTMFSender" missing
FAIL RTCDTMFSender interface object length assert_own_property: self does not have own property "RTCDTMFSender" expected property "RTCDTMFSender" missing
FAIL RTCDTMFSender interface object name assert_own_property: self does not have own property "RTCDTMFSender" expected property "RTCDTMFSender" missing
FAIL RTCDTMFSender interface: existence and properties of interface prototype object assert_own_property: self does not have own property "RTCDTMFSender" expected property "RTCDTMFSender" missing
FAIL RTCDTMFSender interface: existence and properties of interface prototype object's "constructor" property assert_own_property: self does not have own property "RTCDTMFSender" expected property "RTCDTMFSender" missing
FAIL RTCDTMFSender interface: operation insertDTMF(DOMString,unsigned long,unsigned long) assert_own_property: self does not have own property "RTCDTMFSender" expected property "RTCDTMFSender" missing
FAIL RTCDTMFSender interface: attribute ontonechange assert_own_property: self does not have own property "RTCDTMFSender" expected property "RTCDTMFSender" missing
FAIL RTCDTMFSender interface: attribute toneBuffer assert_own_property: self does not have own property "RTCDTMFSender" expected property "RTCDTMFSender" missing
FAIL RTCDTMFToneChangeEvent interface: existence and properties of interface object assert_own_property: self does not have own property "RTCDTMFToneChangeEvent" expected property "RTCDTMFToneChangeEvent" missing
FAIL RTCDTMFToneChangeEvent interface object length assert_own_property: self does not have own property "RTCDTMFToneChangeEvent" expected property "RTCDTMFToneChangeEvent" missing
FAIL RTCDTMFToneChangeEvent interface object name assert_own_property: self does not have own property "RTCDTMFToneChangeEvent" expected property "RTCDTMFToneChangeEvent" missing
FAIL RTCDTMFToneChangeEvent interface: existence and properties of interface prototype object assert_own_property: self does not have own property "RTCDTMFToneChangeEvent" expected property "RTCDTMFToneChangeEvent" missing
FAIL RTCDTMFToneChangeEvent interface: existence and properties of interface prototype object's "constructor" property assert_own_property: self does not have own property "RTCDTMFToneChangeEvent" expected property "RTCDTMFToneChangeEvent" missing
FAIL RTCDTMFToneChangeEvent interface: attribute tone assert_own_property: self does not have own property "RTCDTMFToneChangeEvent" expected property "RTCDTMFToneChangeEvent" missing
PASS RTCStatsReport interface: existence and properties of interface object 
PASS RTCStatsReport interface object length 
PASS RTCStatsReport interface object name 
PASS RTCStatsReport interface: existence and properties of interface prototype object 
PASS RTCStatsReport interface: existence and properties of interface prototype object's "constructor" property 
PASS RTCIdentityProviderGlobalScope interface: existence and properties of interface object 
PASS RTCIdentityProviderRegistrar interface: existence and properties of interface object 
FAIL RTCIdentityAssertion interface: existence and properties of interface object assert_own_property: self does not have own property "RTCIdentityAssertion" expected property "RTCIdentityAssertion" missing
FAIL RTCIdentityAssertion interface object length assert_own_property: self does not have own property "RTCIdentityAssertion" expected property "RTCIdentityAssertion" missing
FAIL RTCIdentityAssertion interface object name assert_own_property: self does not have own property "RTCIdentityAssertion" expected property "RTCIdentityAssertion" missing
FAIL RTCIdentityAssertion interface: existence and properties of interface prototype object assert_own_property: self does not have own property "RTCIdentityAssertion" expected property "RTCIdentityAssertion" missing
FAIL RTCIdentityAssertion interface: existence and properties of interface prototype object's "constructor" property assert_own_property: self does not have own property "RTCIdentityAssertion" expected property "RTCIdentityAssertion" missing
FAIL RTCIdentityAssertion interface: attribute idp assert_own_property: self does not have own property "RTCIdentityAssertion" expected property "RTCIdentityAssertion" missing
FAIL RTCIdentityAssertion interface: attribute name assert_own_property: self does not have own property "RTCIdentityAssertion" expected property "RTCIdentityAssertion" missing
FAIL RTCErrorEvent interface: existence and properties of interface object assert_own_property: self does not have own property "RTCErrorEvent" expected property "RTCErrorEvent" missing
FAIL RTCErrorEvent interface object length assert_own_property: self does not have own property "RTCErrorEvent" expected property "RTCErrorEvent" missing
FAIL RTCErrorEvent interface object name assert_own_property: self does not have own property "RTCErrorEvent" expected property "RTCErrorEvent" missing
FAIL RTCErrorEvent interface: existence and properties of interface prototype object assert_own_property: self does not have own property "RTCErrorEvent" expected property "RTCErrorEvent" missing
FAIL RTCErrorEvent interface: existence and properties of interface prototype object's "constructor" property assert_own_property: self does not have own property "RTCErrorEvent" expected property "RTCErrorEvent" missing
FAIL RTCErrorEvent interface: attribute error assert_own_property: self does not have own property "RTCErrorEvent" expected property "RTCErrorEvent" missing
